Robustify plan_key detection/creation
Currently, Bamboo is missing a fairly significant built-in plan
variable: buildPlanKey, which is technically what's needed to easily
access the artifacts tree via REST.

While its easy to create this variable on a per-plan basis, or as a
Bamboo plugin, I've opted for a cheaper easier solution: generation from
buildKey.

This update _uses_ buildPlanKey if available, but otherwise simply
truncates the buildKey at the last '-' and uses that.